+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - ... <!-- Add new lines here. Version number will be decided later -->
 - feat: Add prop `placeholderDisabled` to `Selectbox`
 - fix: `SelectboxOption.disabled` was ignored
+- fix: Relax typing of props.children, allow nullable `ReactNode`s — even when
+  children are required
 
 ## 0.4.10 — 0.4.11